run: move context to run module
