+++ run.c
@@ -17,7 +17,6 @@
  */
 
 #include <config.h>
-#include "bwrap-oci.h"
 #include <unistd.h>
 #include <stdlib.h>
 #include <error.h>
